pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "webscraper"
version = "0.1.0"
authors = [
    { name = "Your Name" },
]
description = "A web scraper for Glassdoor and Rightmove listings"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "requests>=2.31.0",
    "selenium>=4.0.0",
    "toml>=0.10.2",
    "notion-client>=2.2.1",
    "python-dotenv>=1.0.0",
    "colorama>=0.4.6",
]

[project.scripts]
webscraper = "webscraper.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/webscraper"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = ["-v", "-s", "--tb=short"]
filterwarnings = ["ignore::DeprecationWarning"] 